% File sprsmdl/man/sprsmdl-internal.Rd
% Part of the sprsmdl package for R
% Copyright 2013 Hiroshi Saito
% Distributed under GPL 2 or later

\name{sprsmdl-internal}
\alias{sigmoid}
\title{
  Internal sprsmdl functions
}
\description{
  Internal sprsmdl functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }
