\name{sprsmdl-package}
\alias{sprsmdl}
\docType{package}
\title{
  Sparse modeling toolkit
}
\description{
  R functions to mine sparse models from data.
}
\details{
  Providing functions to find few but significant variables that can
  explain data well.
}
\author{
  Hiroshi Saito \email{die.saito@gmail.com}
}
\keyword{ models }
